+++ interface/app/$libraryId/Layout/index.tsx
@@ -1,6 +1,5 @@
 import clsx from 'clsx';
-import { Suspense } from 'react';
-import { useRef } from 'react';
+import { Suspense, useRef } from 'react';
 import { Navigate, Outlet, useLocation, useParams } from 'react-router-dom';
 import {
 	ClientContextProvider,
@@ -12,8 +11,6 @@ import {
 import { useOperatingSystem } from '~/hooks/useOperatingSystem';
 import { usePlatform } from '~/util/Platform';
 import { QuickPreview } from '../Explorer/QuickPreview';
-import { TopBarContext } from '../PageLayout';
-import TopBar from '../TopBar/.';
 import Sidebar from './Sidebar';
 import Toasts from './Toasts';
 
@@ -21,11 +18,11 @@ const Layout = () => {
 	const { libraries, library } = useClientContext();
 
 	const os = useOperatingSystem();
-	const topBarChildrenRef = useRef<HTMLDivElement>(null);
 
 	initPlausible({
 		platformType: usePlatform().platform === 'tauri' ? 'desktop' : 'web'
 	});
+
 	usePlausiblePageViewMonitor({ currentPath: useLocation().pathname });
 
 	if (library === null && libraries.data) {
