+++ interface/app/$libraryId/index.tsx
@@ -1,24 +1,34 @@
 import { RouteObject } from 'react-router-dom';
 import settingsRoutes from './settings';
 
+// Routes that should be contained within the standard Page layout
+const pageRoutes: RouteObject = {
+	lazy: () => import('./PageLayout'),
+	children: [
+		{
+			path: 'overview',
+			lazy: () => import('./overview')
+		},
+		{ path: 'people', lazy: () => import('./people') },
+		{ path: 'media', lazy: () => import('./media') },
+		{ path: 'spaces', lazy: () => import('./spaces') },
+		{ path: 'debug', lazy: () => import('./debug') },
+		{ path: 'spacedrop', lazy: () => import('./spacedrop') },
+		{ path: 'sync', lazy: () => import('./sync') }
+	]
+};
+// Routes that should render with the top bar - pretty much everything except
+// 404 and settings
+const topBarRoutes: RouteObject = {
+	lazy: () => import('./TopBar/Layout'),
+	children: [
+		pageRoutes,
+		{ path: 'location/:id', lazy: () => import('./location/$id') },
+		{ path: 'tag/:id', lazy: () => import('./tag/$id') }
+	]
+};
 export default [
-	{
-		lazy: () => import('./PageLayout'),
-		children: [
-			{
-				path: 'overview',
-				lazy: () => import('./overview')
-			},
-			{ path: 'people', lazy: () => import('./people') },
-			{ path: 'media', lazy: () => import('./media') },
-			{ path: 'spaces', lazy: () => import('./spaces') },
-			{ path: 'debug', lazy: () => import('./debug') },
-			{ path: 'spacedrop', lazy: () => import('./spacedrop') },
-			{ path: 'sync', lazy: () => import('./sync') }
-		]
-	},
-	{ path: 'location/:id', lazy: () => import('./location/$id') },
-	{ path: 'tag/:id', lazy: () => import('./tag/$id') },
+	topBarRoutes,
 	{
 		path: 'settings',
 		lazy: () => import('./settings/Layout'),
